Factored ToolSocket out to separate files
